refactor: rename primitives to types
