+++ src/blockchain/Blockchain.ts
@@ -16,7 +16,7 @@ import { factory as LoggerFactory } from '../config/ConfigLog'
 import { ERROR_UNKNOWN, ExtrinsicError } from '../errorhandling/ExtrinsicError'
 import Identity from '../identity/Identity'
 import { TxStatus } from './TxStatus'
-import IPublicIdentity from '../primitives/PublicIdentity'
+import IPublicIdentity from '../types/PublicIdentity'
 
 const log = LoggerFactory.getLogger('Blockchain')
