+++ src/delegation/DelegationDecoder.ts
@@ -9,7 +9,7 @@ import { QueryResult } from '../blockchain/Blockchain'
 import { coToUInt8 } from '../crypto/Crypto'
 import { DelegationNode } from './DelegationNode'
 import { DelegationRootNode } from './DelegationRootNode'
-import { Permission } from '../primitives/Delegation'
+import { Permission } from '../types/Delegation'
 
 export type CodecWithId = {
   id: string
